requirements.txt
matplotlib==3.4.1
numpy==1.20.2
scipy==1.6.2
tqdm==4.49.0
langid==1.1.6
pyyaml==5.4.1
torch==1.9.0
transformers==4.8.1
sentencepiece==0.1.96
xxhash==2.0.2
psycopg2==2.9.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gurobipy==9.1.2
